# Catalogo_de_Servicios_Integracion_NIS_Swagger_v3 (2).yaml
openapi: 3.0.3
info:
  title: Consultant Service API - Driving Exam Management System
  version: 3.0.0
  description: |
    REST API for managing CALE classrooms, appointments and driving exam results.

    ## Endpoints
    | Method | Path | Description |
    |--------|------|-------------|
    | GET    | /consultant-service/v1/health | Service health check |
    | POST   | /consultant-service/v1/auth/login | Authenticate and obtain access token |
    | GET    | /consultant-service/v1/cales | List all registered CALE classrooms |
    | GET    | /consultant-service/v1/appointment/{caleId} | Get all applicants assigned to a CALE for today |
    | POST   | /consultant-service/v1/appointment/test-result | Submit exam result (webhook) |
    | GET    | /consultant-service/v1/resource-sync/status | Check if a database record has been modified |

    ## Standard response envelope
    Every endpoint returns responses wrapped in the following envelope:
    ```json
    {
      "success": true,
      "message": "Human-readable description of the operation result",
      "data": {}
    }
    ```
    - `success` — `true` when the operation succeeded, `false` on errors
    - `message` — Informational text about the result of the request
    - `data` — Payload of the response. **Always present, even if empty `{}`**

    ## Authentication
    All endpoints except `/health` and `/auth/login` require a Bearer token:
    ```
    Authorization: Bearer <accessToken>
    ```
...

...

...

...

...

...
    ```
...
    ```

...

...
servers:
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from POST /consultant-service/v1/auth/login

  # ───────────────────────────────────────────────────────────────────────────
  # REUSABLE SCHEMAS
  # ───────────────────────────────────────────────────────────────────────────
  schemas:
...
    ApiResponse:
      type: object
      required:
        - success
        - message
        - data
      properties:
        success:
          type: boolean
          description: true when the operation succeeded, false on any error
          example: true
        message:
          type: string
          description: Human-readable description of the operation result
          example: "Request processed successfully."
        data:
          type: object
          description: Payload of the response. Always present, even if empty.
          additionalProperties: true

    # ── CALE summary (used in GET /cales) ─────────────────────────────────
    CaleSummary:
      type: object
      required:
        - caleId
        - caleName
      properties:
        caleId:
          type: string
          description: Unique identifier of the CALE within its evaluation center.
          example: "CALE-BOG-001"
        caleName:
          type: string
          description: Human-readable name of the CALE classroom
          example: "CALE Bogotá Norte"

    # ── Applicant appointment (used in GET /appointment/{caleId}) ─────────
    ApplicantAppointment:
      type: object
      required:
        - userId
        - firstName
        - lastName
        - documentType
        - documentNumber
        - appointmentId
        - scheduledTime
        - testType
        - licenseCategory
        - status
        - roomId
        - roomName
      properties:
        userId:
          type: string
          description: Unique identifier of the applicant
          example: "USR-001"
        firstName:
          type: string
          description: Applicant's first name
          example: "Juan"
        lastName:
          type: string
          description: Applicant's last name(s)
          example: "Pérez García"
        documentType:
          type: string
          description: |
            Type of identification document:
            - **TI**: Tarjeta de Identidad
            - **CC**: Cédula de Ciudadanía
            - **CE**: Cédula de Extranjería
            - **PP**: Pasaporte
            - **CD**: Carnet Diplomático
            - **PPT**: Permiso de Protección Temporal
          enum: [TI, CC, CE, PP, CD, PPT]
          example: "CC"
        documentNumber:
          type: string
          description: Identification document number
          example: "1234567890"
        email:
          type: string
          format: email
          description: Applicant's email address
          example: "juan.perez@example.com"
        appointmentId:
          type: string
          description: Unique identifier of the scheduled appointment
          example: "APT-2024-001"
        photoUrl:
          type: string
          nullable: true
          description: URL of the applicant's profile photo. Null if not available.
          example: "https://i.pravatar.cc/150?img=1"
        scheduledTime:
          type: string
          description: Appointment time in HH:mm format (24h)
          pattern: '^([01]\d|2[0-3]):[0-5]\d$'
          example: "09:00"
        testType:
          type: string
          description: Type of exam scheduled
          enum: [TEORICO, DESTREZA_INDIVIDUAL, VIA_PUBLICA]
          example: "TEORICO"
        licenseCategory:
          type: string
          description: |
            License category the applicant is applying for:
            - **A1, A2**: Motorcycle categories
            - **B1, B2, B3**: Light vehicle categories
            - **C1, C2, C3**: Heavy vehicle categories
          enum: [A1, A2, B1, B2, B3, C1, C2, C3]
          example: "B1"
        status:
          type: string
          description: Current status of the appointment
          enum: [confirmed, pending, cancelled, completed]
          example: "confirmed"
        roomId:
          type: string
          description: |
            Unique identifier of the exam room within the CALE where the applicant
            will take the exam. Rooms are associated to CALEs in the internal database.
          example: "ROOM-BOG-001-A"
        roomName:
          type: string
          description: Human-readable name of the exam room
          example: "Sala A"

    # ── Module result (percentage-based) ──────────────────────────────────
...
security:
  - BearerAuth: []

# ─────────────────────────────────────────────────────────────────────────────
# PATHS
# ─────────────────────────────────────────────────────────────────────────────
paths:
...
  /consultant-service/v1/auth/login:
    post:
      tags:
        - Auth
      summary: Authenticate and obtain access token
      description: |
        Receives username and password. If valid, returns a JWT access token
        to be used in all subsequent API calls.

        **Token usage:**
        ```
        Authorization: Bearer <accessToken>
        ```
        # NOTE: Token lifetime is fixed at 24 hours (86400 seconds).
        # After expiration the client must re-authenticate.
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - password
              properties:
                username:
                  type: string
                  description: Registered username (typically an email address)
                  example: "operator.cale@example.com"
                password:
                  type: string
                  format: password
                  description: User password
                  example: "S3cur3P@ss!"
            examples:
              validCredentials:
                summary: Valid login credentials
                value:
                  username: "operator.cale@example.com"
                  password: "S3cur3P@ss!"
      responses:
        '200':
          description: Authentication successful — JWT token returned
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        required:
                          - accessToken
                          - tokenType
                        properties:
                          accessToken:
                            type: string
                            # Token lifetime: 24 hours (86400 s). Re-authenticate after expiration.
                            description: JWT access token. Valid for 24 hours.
                            example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOiJVU1ItMDAxIn0.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c"
                          tokenType:
                            type: string
                            enum: [Bearer]
                            description: Token scheme — always "Bearer"
                            example: "Bearer"
              examples:
                loginSuccess:
                  summary: Successful login
                  value:
                    success: true
                    message: "Authentication successful."
                    data:
                      accessToken: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOiJVU1ItMDAxIn0.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c"
                      tokenType: "Bearer"
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
              examples:
                invalidCredentials:
                  summary: Wrong username or password
                  value:
                    success: false
                    message: "The username or password provided is incorrect."
                    data: {}

  # ===========================================================================
  # CALE — List all CALE classrooms
  # ===========================================================================
  /consultant-service/v1/cales:
    get:
      tags:
        - CALE
      summary: Get all registered CALE classrooms
      description: |
        Returns all CALE classrooms registered on the platform.
        **No parameters required.**

        Each item contains only:
        - `caleId` — unique identifier of the CALE (unique per center)
        - `caleName` — human-readable name

        Use the returned `caleId` values to query applicant assignments via
        `GET /appointment/{caleId}`.
      operationId: getAllCales
      responses:
        '200':
          description: List of registered CALE classrooms
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        required:
                          - cales
                        properties:
                          cales:
                            type: array
                            items:
                              $ref: '#/components/schemas/CaleSummary'
              examples:
                caleList:
                  summary: Registered CALE classrooms
                  value:
                    success: true
                    message: "CALE classrooms retrieved successfully."
                    data:
                      cales:
                        - caleId: "CALE-BOG-001"
                          caleName: "CALE Bogotá Norte"
                        - caleId: "CALE-BOG-002"
                          caleName: "CALE Bogotá Sur"
                        - caleId: "CALE-MED-001"
                          caleName: "CALE Medellín"
        '401':
          description: Missing or invalid authentication token
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
              examples:
                unauthorized:
                  summary: No valid token provided
                  value:
                    success: false
                    message: "A valid Bearer token is required to access this resource."
                    data: {}
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
              examples:
                internalError:
                  summary: Unexpected server error
                  value:
                    success: false
                    message: "An unexpected error occurred. Please try again later."
                    data: {}

  # ===========================================================================
  # APPOINTMENTS — Applicants by CALE
  # ===========================================================================
  /consultant-service/v1/appointment/{caleId}:
    get:
      tags:
        - Appointments
      summary: Get all applicants assigned to a CALE for today
      description: |
        Returns the complete list of applicants with appointments scheduled for the
        **current day** in the CALE identified by `caleId`.

        Each applicant includes their personal data, appointment details and the
        **exam room** (`roomId`, `roomName`) within the CALE where they will take
        the exam. Rooms are linked to CALEs in the internal database.

        > All examples use `testType: TEORICO`.
      operationId: getAppointmentsByCaleId
      parameters:
        - name: caleId
          in: path
          required: true
          description: |
            Unique identifier of the CALE classroom.
            Obtain valid values from `GET /consultant-service/v1/cales`.
          schema:
            type: string
          examples:
            bogotaNorth:
              value: "CALE-BOG-001"
              summary: CALE Bogotá Norte
            bogotaSouth:
              value: "CALE-BOG-002"
              summary: CALE Bogotá Sur
      responses:
        '200':
          description: Full list of applicants assigned to the CALE for today
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        required:
                          - appointmentDate
                          - totalApplicants
                          - applicants
                        properties:
                          appointmentDate:
                            type: string
                            format: date
                            description: Current date for which appointments are listed
                            example: "2024-02-17"
                          totalApplicants:
                            type: integer
                            description: Total number of applicants with appointments today in this CALE
                            example: 3
                          applicants:
                            type: array
                            description: |
                              Complete list of applicants assigned to this CALE today.
                              Each applicant includes the exam room (roomId, roomName)
                              within the CALE where they will take the exam.
                            items:
                              $ref: '#/components/schemas/ApplicantAppointment'
              examples:
                caleWithApplicants:
                  summary: CALE with multiple applicants scheduled today
                  value:
                    success: true
                    message: "Applicants for CALE-BOG-001 retrieved successfully."
                    data:
                      appointmentDate: "2024-02-17"
                      totalApplicants: 3
                      applicants:
                        - userId: "USR-001"
                          firstName: "Juan"
                          lastName: "Pérez García"
                          documentType: "CC"
                          documentNumber: "1234567890"
                          email: "juan.perez@example.com"
                          appointmentId: "APT-2024-001"
                          photoUrl: "https://i.pravatar.cc/150?img=1"
                          scheduledTime: "09:00"
                          testType: "TEORICO"
                          licenseCategory: "B1"
                          status: "confirmed"
                          roomId: "ROOM-BOG-001-A"
                          roomName: "Sala A"
                        - userId: "USR-002"
                          firstName: "María"
                          lastName: "López Rodríguez"
                          documentType: "CE"
                          documentNumber: "9876543210"
                          email: "maria.lopez@example.com"
                          appointmentId: "APT-2024-002"
                          photoUrl: null
                          scheduledTime: "10:00"
                          testType: "TEORICO"
                          licenseCategory: "A2"
                          status: "confirmed"
                          roomId: "ROOM-BOG-001-A"
                          roomName: "Sala A"
                        - userId: "USR-003"
                          firstName: "Carlos"
                          lastName: "González Martínez"
                          documentType: "TI"
                          documentNumber: "5555666677"
                          email: "carlos.gonzalez@example.com"
                          appointmentId: "APT-2024-003"
                          photoUrl: "https://i.pravatar.cc/150?img=3"
                          scheduledTime: "11:00"
                          testType: "TEORICO"
                          licenseCategory: "C1"
                          status: "pending"
                          roomId: "ROOM-BOG-001-B"
                          roomName: "Sala B"
                emptyCale:
                  summary: CALE with no appointments today
                  value:
                    success: true
                    message: "No applicants assigned to CALE-MED-001 for today."
                    data:
                      appointmentDate: "2024-02-17"
                      totalApplicants: 0
                      applicants: []
        '400':
          description: Invalid caleId format
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
              examples:
                invalidFormat:
                  summary: caleId with invalid format
                  value:
                    success: false
                    message: "The provided caleId has an invalid format."
                    data:
                      field: "caleId"
                      providedValue: "!!!INVALID"
        '401':
          description: Missing or invalid authentication token
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
              examples:
                unauthorized:
                  summary: No valid token provided
                  value:
                    success: false
                    message: "A valid Bearer token is required to access this resource."
                    data: {}
        '404':
          description: CALE not found
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
              examples:
                caleNotFound:
                  summary: No CALE registered with that caleId
                  value:
                    success: false
                    message: "No CALE was found with caleId: CALE-XXX-999"
                    data:
                      caleId: "CALE-XXX-999"
                      suggestion: "Use GET /consultant-service/v1/cales to retrieve valid caleId values."
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
              examples:
                internalError:
                  summary: Unexpected server error
                  value:
                    success: false
                    message: "An unexpected error occurred. Please try again later."
                    data: {}

  # ===========================================================================
  # TEST RESULTS — Webhook
  # ===========================================================================
...
  /consultant-service/v1/resource-sync/status:
    get:
      tags:
        - Resource Sync
      summary: Check whether a database record has been modified
      description: |
        **Purely informational endpoint.** No data is created or modified.

        Allows an external consumer to query whether a specific database record
        (identified by `tableName` + `recordId`) has been updated since the last
        time the consumer processed it.

        ## How to use
        1. Call this endpoint providing the table name and the record identifier.
        2. If `hasChanged = true` → the consumer should trigger its own data
           refresh / sync process against the database.
        3. If `hasChanged = false` → no action needed; local data is still valid.

        This endpoint does **not** return the actual changed data. It only signals
        **what changed** (field names) and **when** (`lastModifiedAt`), so the
        consumer can decide whether to execute its update process.

        ## Query parameters
        | Parameter | Required | Description |
        |-----------|----------|-------------|
        | `tableName` | ✅ | Name of the database table to check (e.g., `appointments`, `cales`, `applicants`) |
        | `recordId` | ✅ | Primary key or unique identifier of the specific record to check |
      operationId: getResourceSyncStatus
      parameters:
        - name: tableName
          in: query
          required: true
          description: |
            Name of the database table containing the record to check.
            Examples: `appointments`, `cales`, `applicants`, `rooms`
          schema:
            type: string
          example: "appointments"
        - name: recordId
          in: query
          required: true
          description: |
            Primary key or unique identifier of the specific record to check.
            Must correspond to a record in the given `tableName`.
          schema:
            type: string
          example: "APT-2024-001"
      responses:
        '200':
          description: Sync status for the requested record returned successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        required:
                          - tableName
                          - recordId
                          - hasChanged
                          - lastModifiedAt
                        properties:
                          tableName:
                            type: string
                            description: The database table that was checked
                            example: "appointments"
                          recordId:
                            type: string
                            description: The specific record identifier that was checked
                            example: "APT-2024-001"
                          hasChanged:
                            type: boolean
                            description: |
                              true  = The record was modified. The consumer should
                                      trigger its data refresh process.
                              false = No changes detected. No action needed.
                            example: true
                          lastModifiedAt:
                            type: string
                            format: date-time
                            description: Timestamp of the last modification to this record
                            example: "2024-02-17T08:45:00Z"
                          modifiedFields:
                            type: array
                            nullable: true
                            description: |
                              List of field names that were modified in this record.
                              Null when hasChanged is false.
                            items:
                              type: string
                            example: ["status", "scheduledTime"]
              examples:
                recordChanged:
                  summary: Record was modified — consumer should refresh
                  value:
                    success: true
                    message: "The record has been modified since the last known state."
                    data:
                      tableName: "appointments"
                      recordId: "APT-2024-001"
                      hasChanged: true
                      lastModifiedAt: "2024-02-17T08:45:00Z"
                      modifiedFields: ["status", "scheduledTime"]
                recordUnchanged:
                  summary: Record has not changed — no action needed
                  value:
                    success: true
                    message: "No changes detected for this record."
                    data:
                      tableName: "appointments"
                      recordId: "APT-2024-001"
                      hasChanged: false
                      lastModifiedAt: "2024-02-16T07:00:00Z"
                      modifiedFields: null
                caleRecordChanged:
                  summary: CALE record was modified
                  value:
                    success: true
                    message: "The record has been modified since the last known state."
                    data:
                      tableName: "cales"
                      recordId: "CALE-BOG-001"
                      hasChanged: true
                      lastModifiedAt: "2024-02-16T18:30:00Z"
                      modifiedFields: ["caleName"]
        '400':
          description: Missing or invalid query parameters
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
              examples:
                missingParams:
                  summary: Required query parameters not provided
                  value:
                    success: false
                    message: "Both 'tableName' and 'recordId' query parameters are required."
                    data:
                      missingParams: ["recordId"]
        '401':
          description: Missing or invalid authentication token
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
              examples:
                unauthorized:
                  summary: No valid token provided
                  value:
                    success: false
                    message: "A valid Bearer token is required to access this resource."
                    data: {}
        '404':
          description: Record not found
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
              examples:
                recordNotFound:
                  summary: No record found for the given tableName + recordId
                  value:
                    success: false
                    message: "No record was found in table 'appointments' with recordId: APT-9999"
                    data:
                      tableName: "appointments"
                      recordId: "APT-9999"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
              examples:
                internalError:
                  summary: Unexpected server error
                  value:
                    success: false
                    message: "An unexpected error occurred. Please try again later."
                    data: {}